% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{centers.hclust}
\alias{centers.hclust}
\alias{mean.values.clusters}
\title{centers.hclust}
\usage{
centers.hclust(bases.hclust, bases, s)
}
\arguments{
\item{bases.hclust}{is an hclust objet}

\item{bases}{is a dats frame with segmentation bases used in the hclust object}
}
\value{
centros.bases  a matrix with centers
}
\description{
This function comnputs the mean value of bases in the segments
we want to form later with k-means.
}

