% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.functions.R
\name{compute.importance.for.one}
\alias{compute.importance.for.one}
\title{computes de importance of attributes for one individual}
\usage{
compute.importance.for.one(conjoint.results)
}
\arguments{
\item{conjoint.results}{a list with the results of a conjoint analysis}
}
\value{
conjoint.results       returns the conjoint.results list modified
}
\description{
This function computes attributes' importance for one individual. It takes the results of a conjoint analysis,
the list conjoint.results, and then I compute the the importance of every attribute. It uses
sum zero constraints to estimate part worths. I adapt  code from Marketing Data Science book, chapter 1.
}
\examples{
# not run This is a function called by importance.of.attributes()
#data(MDSConjointData)
#names(MDSConjointData$osc)
#osc<-MDSConjointData$osc
#osc.imp<-importance.of.attributes(osc$ratings, osc$bundles, osc$design)
}

