% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/part.worths.R
\name{conjoint.estimation}
\alias{conjoint.estimation}
\title{conjoint.estimation}
\usage{
conjoint.estimation(ratings, bundles, design.l, rank = 0,
  conj.contrasts = "sum")
}
\arguments{
\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{design.l}{a list with the conjoint design (attributes and levels)}

\item{rank}{if rank==1, then transform a ranking into utilities}

\item{conj.contrasts}{if conj.contrasts=="sum", then  options(contrasts = c("contr.sum", "contr.poly")), otherwise, treatment}
}
\value{
conjoint.results.all    the object returned by the function
}
\description{
This function computes the partworth of attribute levels for each
respondent. Uses treatment effects; ratings is a data frame of clients
(rows) by bundle rated (colums); bundles is a data frame of
bundles (rows) by attributes (colums)
}
\examples{

data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
class(osc)
names(osc)
osc.conjoint <- conjoint.estimation(osc$ratings, osc$bundles, osc$design)
names(osc.conjoint)
# [1] "summary"     "fit"         "part.worths"  "prediction"
head(osc.conjoint$summary)
head(osc.conjoint$fit)
head(osc.conjoint$part.worths)
head(osc.conjoint$prediction)
}

