% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.prediction.pw.R
\name{cupi}
\alias{cupi}
\title{This function computes for client i the utility of each product profile from a data frame of part worths}
\usage{
cupi(pw, mp, n.attributes)
}
\arguments{
\item{pw}{a data frame with all clients' partworths}

\item{mp}{a data frame with the description of competitors' existing product profiles}

\item{n.attributes}{the number attributs of conjoint design}
}
\value{
upi     a matrix with the prediction of utilities for  client i
}
\description{
This function computes for client i the utility of each market profile (given rivals' profiles and new profiles
) for  consumer i.  The data frame mp has the data of existing profiles
for which we are going to compute the market shares.
It is a table of market product profiles (rows) by
attributes (columns), wp is a data frame of client i (rows) by part worths
 (colums), design.l is a list with conjoint desing.
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
#this function is called by uop().
osc.conjoint <- conjoint.estimation(osc$ratings, osc$bundles, osc$design)
names(osc.conjoint)
# [1] "summary"     "fit"         "part.worths"  "prediction"
#head(osc.conjoint$summary)
head(osc.conjoint$fit)
head(osc.conjoint$part.worths)
head(osc.conjoint$prediction)
#osc.cupi <-cupi(osc$market.profiles, osc.conjoint$part.worths[1,], osc$design)
#head(osc.cupi)
}

