% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df2list}
\alias{df2list}
\title{df2list}
\usage{
df2list(df)
}
\arguments{
\item{df}{A data frame}
}
\value{
df.list a list
}
\description{
This function transforms a data frame's into a list according to columns
}
\examples{
xy.df <- data.frame(x = runif(10),  y = runif(10))
df2list(xy.df)
}

