% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.prediction.pw.R
\name{fac2char}
\alias{fac2char}
\title{This function transform factor variable into character type}
\usage{
fac2char(mpi)
}
\arguments{
\item{mpi}{a data frame}
}
\value{
mpi a data frame transformed
}
\description{
This functions takes as data frame with variables of factor type and transform them into
a character type. It is usefull for computing product profile utility from a matrix of part worths
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
fac2char(osc$market.profiles)
}

