% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.functions.R
\name{importance.of.attributes}
\alias{importance.of.attributes}
\title{Computes the importance of producy attributes}
\usage{
importance.of.attributes(ratings, bundles, design.l, rank = 0)
}
\arguments{
\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{design.l}{a list with attributs and their levels}

\item{rank}{if rank==1, then transform a ranking into utilities}
}
\value{
conjoint.list          a list with three data frames:
                               1) fit, with the model fited for each individual,
                               2) part.worths, with the estimated part worths for all individuals
                               3) imp, wiht the attributes' importance for each individual
}
\description{
This function computes attributes' importance. First I prepare the
list conjoint.results, after that I compute the sample's mean. It uses
sum zero contrasts to estimate part worths; ratings is a data frame of clients (rows) by bundles
rated (colums); bundles is a data frame of bundles (rows) by attributes
(colums); design.l is a list with the conjoint design (attributes and levels)
Frist we prepare data
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
names(MDSConjointData$osc)
osc<-MDSConjointData$osc
osc.imp<-importance.of.attributes(osc$ratings, osc$bundles, osc$design)
names(osc.imp)
head(osc.imp$fit)
head(osc.imp$part.worths)
head(osc.imp$imp)
names(MDSConjointData$tire)
tire<-MDSConjointData$tire
tire.imp<-importance.of.attributes(tire$ratings, tire$bundles, tire$design)
names(tire.imp)
head(tire.imp$fit)
head(tire.imp$part.worths)
head(tire.imp$imp)
}

