% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.prediction.R
\name{ms.fe.conjoint}
\alias{ms.fe.conjoint}
\title{This function computes the market share of  product profiles}
\usage{
ms.fe.conjoint(market.profiles, ratings, bundles, rank = 0)
}
\arguments{
\item{market.profiles}{a data frame with the description of competitors' existing product profiles}

\item{ratings}{a data frame frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{rank}{if rank==1, then transform a ranking into utilities}
}
\value{
ms
}
\description{
This function computes the market share of  product profiles
according to the first choice rule. Uses treatment effects to estimate part worth utilities of profiles;
the matrix of profiles for which we are going to compute market
profiles is a matrix of existing product profiles in the market (rows)
by attributes (columns); ratings is a matrix of clients (rows) by
bundles rated (colums); bundles is a matrix of bundles (rows)
by attributes (colums)
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
names(osc)
sapply(osc, class)
sapply(osc$market.proiles, class)
ms.fe.conjoint(osc$market.profiles, osc$ratings, osc$bundles)
class(ms.fe.conjoint(osc$market.profiles, osc$ratings, osc$bundles))
}

