% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.prediction.R
\name{ms.logit.conjoint}
\alias{ms.logit.conjoint}
\title{Computes the market share of the product profiles}
\usage{
ms.logit.conjoint(market.profiles, ratings, bundles, rank = 0)
}
\arguments{
\item{market.profiles}{a matrix with the description of competitors' product profiles}

\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{rank}{if rank==1, then transform a ranking into utilities}
}
\value{
ms
}
\description{
This function
computes the market share of the product profiles according to the
logit rule. It uses treatment effects to estimate part worths of profiles;
profiles is a matrix of product profiles in the market (rows)
by attributes (columns); ratings is a matrix of clients (rows) by
bundles rated (colums); bundles is a matrix of bundles (rows)
by attributes (colums)
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<- MDSConjointData$osc
names(osc)
ms.logit.conjoint(osc$market.profiles, osc$ratings, osc$bundles)
class(ms.logit.conjoint(osc$market.profiles, osc$ratings, osc$bundles))
}

