% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.prediction.R
\name{ms.us.conjoint}
\alias{ms.us.conjoint}
\title{This function computes the market share of  product profiles}
\usage{
ms.us.conjoint(market.profiles, ratings, bundles, rank = 0)
}
\arguments{
\item{market.profiles}{a matrix with the description of competitors' product profiles}

\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{rank}{if rank==1, then transform a ranking into utilities}
}
\value{
ms
}
\description{
This function computes the market share of  product profiles
according to the utility share rule (also known as BTL rule). Uses treatment effects for estimating part worths of profiles;
is the data frame of profiles for which we are going to compute market
profiles is a matrix of product profiles in the market (rows)
by attributes (columns); ratings is a matrix of clients (rows) by
bundles rated (colums); bundles is a matrix of bundles (rows)
by attributes (colums).
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
ms.us.conjoint(osc$market.profiles, osc$ratings, osc$bundles)
class(ms.us.conjoint(osc$market.profiles, osc$ratings, osc$bundles))
}

