% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.ms.R
\name{optim.ms.logit}
\alias{optim.ms.logit}
\title{Identifies the product profiles that maximazes  market share, logit rule}
\usage{
optim.ms.logit(ratings, bundles, market.profiles, design, rank = 0, hpb = 0)
}
\arguments{
\item{ratings}{a data frame with all clients' ratings}

\item{bundles}{a data frame with all product profiles rated by clients}

\item{market.profiles}{a data frame with competitiors' profiles}

\item{design}{a list with the experiment description (attributes and levels)}

\item{rank}{if rank==1, then transform a ranking into utilities}

\item{hpb}{if hpb==1, the hide progress bar, for rmarkdown documents}
}
\value{
optim.list             a list with with the optim profile as well as optim market share
}
\description{
Computes market share for actual profiles as well as for all possible profiles
in order to indentify the bundle that maximazes market share given actual profiles
We use first choice rule
If we combine all three functions into one function, we can use rule=1 for first choice, rule=2 for utility share, and
the rest for logit. With a if sentence we hace combine all 3 functions
into one.
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
class(osc)
names(osc)
rat<-osc$ratings
bun<-osc$bundles
mp<-osc$market.profiles
design<-osc$design
osc.ms.op.logit<-optim.ms.logit(rat, bun, mp, design, hpb=1)
osc.ms.op.logit
}
\references{
SAS Institute Inc., SAS*TeclmicalReport R-109, Conjoint Analysis Examples, Gary, NC: SAS Institute Inc., 1993.85 pp.
}
\keyword{market}
\keyword{optim}
\keyword{share}

