% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optim.ms.pw.R
\name{optim.ms.us.pw}
\alias{optim.ms.us.pw}
\title{Identifies the product profiles that maximazes  market share, share of utility rule}
\usage{
optim.ms.us.pw(mp, pw, design, hpb = 0)
}
\arguments{
\item{mp}{a data frame with competitiors' profiles}

\item{pw}{a data frame with all clients' part worths}

\item{design}{a list with the experiment description (attributes and levels)}

\item{hpb}{if hpb==1, then hide the progress bar. This is for rmarkdown}
}
\value{
optim.list             a list with with the optim profile as well as optim market share
}
\description{
Computes market share for actual profiles as well as for all possible profiles
in order to indentify the bundle that maximazes market share given actual profiles
We use utility share rule
If we combine all three functions into one function, we can use rule=1 for first choice, rule=2 for utility share, and
the rest for logit. With a if sentence we hace combine all 3 functions
into one.
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
class(osc)
names(osc)
rat<-osc$ratings
bun<-osc$bundles
mp<-osc$market.profiles
design<-osc$design
osc.ms.op.us<-optim.ms.utility.share(rat, bun, mp, design, hpb=1)
osc.ms.op.us

}
\keyword{market}
\keyword{optim}
\keyword{share}

