% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform2factor}
\alias{transform2factor}
\title{transform2factor}
\usage{
transform2factor(df)
}
\arguments{
\item{df}{A data frame with variables in numeric format}
}
\value{
df a data frame with all variables in factor format
}
\description{
This function transforms a data frame's variables into factors
This function is a convenient way to  quickly prepare data from other packages to
test with this package
}
\examples{
n = c(2, 3, 5)
s = c("aa", "bb", "cc")
b = c(TRUE, FALSE, TRUE)
df = data.frame(n, s, b)
df
class(df$n)
class(df$s)
class(df$b)
sapply(df, class)

df.factor <- transform2factor(df)
class(df$n)
class(df$s)
class(df$b)
}

