% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{transform2utilities}
\alias{transform.ranking.to.utilities}
\alias{transform2utilities}
\title{transform2utilities}
\usage{
transform2utilities(y)
}
\arguments{
\item{y}{A vector of bundles ranked by clients}
}
\value{
y A vector of bundles transformed into utilities
}
\description{
This function transforms a vector of ranked profiles into a vector of utilities
It uses SAS monotonic transformation of preference judgements as described in SAS technical Report R-109, p. 2.
}
\examples{
ranking<-c(1,2,3,4,5,6,7,8,9,10,11,12,13,14,15,16,17,18)
ranking
ratings <- transform2utilities(ranking)
ratings
}
\references{
SAS Institute Inc., SAS Technical Report R-109, Conjoint Analysis Examples, Cary, NC: SAS Institute Inc., 1993.85 pp.
}

