% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ms.prediction.pw.R
\name{uop}
\alias{uop}
\title{This function computes the utility of each product profile from a data frame of part worths}
\usage{
uop(mp, pw, design.l)
}
\arguments{
\item{mp}{a data frame with the description of competitors' existing product profiles}

\item{pw}{a data frame with all clients' partworths}

\item{design.l}{a list with conjoint design}
}
\value{
up     a matrix with the prediction of utilities for all clients
}
\description{
This function computes the utility of each profile (given rivals' profiles and new profiles
) for each consumer.  The data frame mp has the data of existing profiles
for which we are going to compute the market shares.
It is a table of market product profiles (rows) by
attributes (columns), wp is a data frame of clients (rows) by part worths
 (colums), design.l is a list with conjoint desing.
}
\examples{
data(MDSConjointData)
names(MDSConjointData)
osc<-MDSConjointData$osc
#this function is called by market share function.
osc.conjoint <- conjoint.estimation(osc$ratings, osc$bundles, osc$design)
names(osc.conjoint)
# [1] "summary"     "fit"         "part.worths"  "prediction"
#head(osc.conjoint$summary)
head(osc.conjoint$fit)
head(osc.conjoint$part.worths)
head(osc.conjoint$prediction)
osc.uop <-uop(osc$market.profiles, osc.conjoint$part.worths, osc$design)
head(osc.uop)

}

