% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importance.functions.R
\name{visualize.importance}
\alias{visualize.importance}
\title{Plots the mean importance of attributes}
\usage{
visualize.importance(part.worths, importance, design.l)
}
\arguments{
\item{part.worths}{a  data frame with as many rows as clients and as many columns as atributes' levels}

\item{importance}{a  data frame with as many rows as clinetes and as mnay columns as attributes' importance}

\item{design.l}{a list with attributs and their levels}
}
\description{
This functions print attributes' importance using R standard plotting system
}
\examples{
#data(MDSConjointData)
#names(MDSConjointData)
#osc<-MDSConjointData
#osc.imp<-importance.of.attributes(osc$ratings, osc$bundles, osc$design)
#names(osc.imp)
#visualize.importance(osc.imp$part.worths, osc.imp$imp, osc$design)
#tire <- MDSConjointData$tire
#tire.imp<-importance.of.attributes(tire$ratings, tire$bundles, tire$design)
#names(tire.imp)
#visualize.importance(tires.imp$part.worths,tires.imp$imp, tire$design)
}

